--- serverless.yaml
@@ -1,74 +1,0 @@
-service: quiz-app
-
-provider:
-    # for locally we will use dev once we finish and deploy we can change it to aws and add iamRoleStatements
-    name: dev
-    runtime: python3.12
-    region: us-north-1
-    iamRoleStatements:
-        - Effect: Allow
-          Action:
-              - dynamodb:Query
-              - dynamodb:Scan
-              - dynamodb:GetItem
-              - dynamodb:PutItem
-              - dynamodb:UpdateItem
-              - dynamodb:DeleteItem
-          Resource: 'arn:aws:dynamodb:us-north-1:*:table/*'
-
-plugins:
-    # To install this serverless plugin install -n serverless-offline, and then you can run locally using serverless offline (you need to install serverless@3)
-    - serverless-offline
-
-functions:
-    seedData:
-        handler: handler.seed_data
-        events:
-            - http:
-                  path: seed
-                  method: post
-    # Quiz CRUD
-    createQuiz:
-        handler: handler.create_quiz
-        events:
-            - http:
-                  path: quiz
-                  method: post
-    getQuizById:
-        handler: handler.get_quiz_by_id
-        events:
-            - http:
-                  path: quiz/{quizId}
-                  method: get
-    getAllQuizzes:
-        handler: handler.get_all_quizzes
-        events:
-            - http:
-                  path: quiz
-                  method: get
-
-resources:
-    Resources:
-        QuizTable:
-            Type: AWS::DynamoDB::Table
-            Properties:
-                TableName: QuizTable
-                AttributeDefinitions:
-                    - AttributeName: PK
-                      AttributeType: S # Partition Key (Primary Key)
-                    - AttributeName: SK
-                      AttributeType: S # Sort Key
-                KeySchema:
-                    - AttributeName: PK
-                      KeyType: HASH # Partition Key
-                    - AttributeName: SK
-                      KeyType: RANGE # Sort Key
-                ProvisionedThroughput:
-                    ReadCapacityUnits: 5
-                    WriteCapacityUnits: 5
-                # Enable Point-in-time Recovery for safety (Optional)
-                PointInTimeRecoverySpecification:
-                    PointInTimeRecoveryEnabled: true
-                # Table-level Encryption (Optional)
-                SSESpecification:
-                    SSEEnabled: true
